+++ sdk/cognitivelanguage/azure-ai-language-questionanswering/samples/async_samples/sample_query_text_async.py
@@ -1,3 +1,4 @@
+# pylint: disable=line-too-long,useless-suppression
 # coding=utf-8
 # ------------------------------------
 # Copyright (c) Microsoft Corporation.
@@ -36,14 +37,14 @@ async def sample_query_text():
         input = qna.AnswersFromTextOptions(
             question=question,
             text_documents=[
-                "Power and charging. It takes two to four hours to charge the Surface Pro 4 battery fully from an empty state. " +
-                "It can take longer if you’re using your Surface for power-intensive activities like gaming or video streaming while you’re charging it.",
-                "You can use the USB port on your Surface Pro 4 power supply to charge other devices, like a phone, while your Surface charges. " +
-                "The USB port on the power supply is only for charging, not for data transfer. If you want to use a USB device, plug it into the USB port on your Surface.",
-            ]
+                "Power and charging. It takes two to four hours to charge the Surface Pro 4 battery fully from an empty state. "
+                + "It can take longer if you’re using your Surface for power-intensive activities like gaming or video streaming while you’re charging it.",
+                "You can use the USB port on your Surface Pro 4 power supply to charge other devices, like a phone, while your Surface charges. "
+                + "The USB port on the power supply is only for charging, not for data transfer. If you want to use a USB device, plug it into the USB port on your Surface.",
+            ],
         )
 
-        output = await client.get_answers_from_text(input)
+        output = await client.question_answering.get_answers_from_text(input)
         if output.answers:
             best_answer = [a for a in output.answers if a.confidence and a.confidence > 0.9][0]
             print("Q: {}".format(question))
@@ -54,5 +55,5 @@ async def sample_query_text():
     # [END query_text_async]
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     asyncio.run(sample_query_text())