+++ sdk/cognitivelanguage/azure-ai-language-questionanswering/tests/helpers.py
@@ -5,16 +5,17 @@
 # license information.
 # --------------------------------------------------------------------------
 
+
 class QnaAuthoringHelper:
 
     def create_test_project(
         client,
-        project_name = "IssacNewton",
-        is_deployable = False,
-        add_sources = False,
-        get_export_url = False,
-        delete_old_project = False,
-        add_qnas = False,
+        project_name="IssacNewton",
+        is_deployable=False,
+        add_sources=False,
+        get_export_url=False,
+        delete_old_project=False,
+        add_qnas=False,
         **kwargs
     ):
         # create project
@@ -24,10 +25,9 @@ class QnaAuthoringHelper:
                 "description": "biography of Sir Issac Newton",
                 "language": "en",
                 "multilingualResource": True,
-                "settings": {
-                    "defaultAnswer": "no answer"
-                }
-            })
+                "settings": {"defaultAnswer": "no answer"},
+            },
+        )
 
         # add sources
         if is_deployable or add_sources:
@@ -45,8 +45,8 @@ class QnaAuthoringHelper:
                     "value": {
                         "displayName": "Issac Newton Bio",
                         "sourceUri": "https://wikipedia.org/wiki/Isaac_Newton",
-                        "sourceKind": "url"
-                    }
+                        "sourceKind": "url",
+                    },
                 }
             ],
             **kwargs
@@ -55,19 +55,12 @@ class QnaAuthoringHelper:
 
     def export_project(client, project_name, delete_project=True, **kwargs):
         # export project
-        export_poller = client.begin_export(
-            project_name=project_name,
-            file_format="json",
-            **kwargs
-        )
+        export_poller = client.begin_export(project_name=project_name, file_format="json", **kwargs)
         result = export_poller.result()
 
         # delete old project
         if delete_project:
-            delete_poller = client.begin_delete_project(
-                project_name=project_name,
-                **kwargs
-            )
+            delete_poller = client.begin_delete_project(project_name=project_name, **kwargs)
             delete_poller.result()
         return result["resultUrl"]
 
@@ -76,12 +69,12 @@ class QnaAuthoringAsyncHelper:
 
     async def create_test_project(
         client,
-        project_name = "IssacNewton",
-        is_deployable = False,
-        add_sources = False,
-        get_export_url = False,
-        delete_old_project = False,
-        add_qnas = False,
+        project_name="IssacNewton",
+        is_deployable=False,
+        add_sources=False,
+        get_export_url=False,
+        delete_old_project=False,
+        add_qnas=False,
         **kwargs
     ):
         # create project
@@ -91,17 +84,18 @@ class QnaAuthoringAsyncHelper:
                 "description": "biography of Sir Issac Newton",
                 "language": "en",
                 "multilingualResource": True,
-                "settings": {
-                    "defaultAnswer": "no answer"
-                }
-            })
+                "settings": {"defaultAnswer": "no answer"},
+            },
+        )
 
         # add sources
         if is_deployable or add_sources:
             await QnaAuthoringAsyncHelper.add_sources(client, project_name, **kwargs)
 
         if get_export_url:
-            return await QnaAuthoringAsyncHelper.export_project(client, project_name, delete_project=delete_old_project, **kwargs)
+            return await QnaAuthoringAsyncHelper.export_project(
+                client, project_name, delete_project=delete_old_project, **kwargs
+            )
 
     async def add_sources(client, project_name, **kwargs):
         update_sources_poller = await client.begin_update_sources(
@@ -112,8 +106,8 @@ class QnaAuthoringAsyncHelper:
                     "value": {
                         "displayName": "Issac Newton Bio",
                         "sourceUri": "https://wikipedia.org/wiki/Isaac_Newton",
-                        "sourceKind": "url"
-                    }
+                        "sourceKind": "url",
+                    },
                 }
             ],
             **kwargs
@@ -122,18 +116,11 @@ class QnaAuthoringAsyncHelper:
 
     async def export_project(client, project_name, delete_project=True, **kwargs):
         # export project
-        export_poller = await client.begin_export(
-            project_name=project_name,
-            file_format="json",
-            **kwargs
-        )
+        export_poller = await client.begin_export(project_name=project_name, file_format="json", **kwargs)
         result = await export_poller.result()
 
         # delete old project
         if delete_project:
-            delete_poller = await client.begin_delete_project(
-                project_name=project_name,
-                **kwargs
-            )
+            delete_poller = await client.begin_delete_project(project_name=project_name, **kwargs)
             await delete_poller.result()
         return result["resultUrl"]