+++ sdk/cognitivelanguage/azure-ai-language-questionanswering/tests/test_create_and_deploy_project.py
@@ -4,6 +4,7 @@
 # Licensed under the MIT License.
 # ------------------------------------
 import pytest
+import os
 
 from azure.ai.language.questionanswering.authoring import AuthoringClient
 from azure.core.credentials import AzureKeyCredential
@@ -19,9 +20,15 @@ class TestCreateAndDeploy(QuestionAnsweringTestCase):
         client = AuthoringClient(qna_creds["qna_endpoint"], AzureKeyCredential(qna_creds["qna_key"]))
         assert client._config.polling_interval == 5
         # test override
-        client = AuthoringClient(qna_creds["qna_endpoint"], AzureKeyCredential(qna_creds["qna_key"]), polling_interval=1)
+        client = AuthoringClient(
+            qna_creds["qna_endpoint"], AzureKeyCredential(qna_creds["qna_key"]), polling_interval=1
+        )
         assert client._config.polling_interval == 1
 
+    @pytest.mark.skipif(
+        not all(os.getenv(v) for v in ["AZURE_TENANT_ID", "AZURE_CLIENT_ID", "AZURE_CLIENT_SECRET"]),
+        reason="Missing AAD env vars for DefaultAzureCredential",
+    )
     def test_create_project_aad(self, recorded_test, qna_creds):
         token = self.get_credential(AuthoringClient)
         client = AuthoringClient(qna_creds["qna_endpoint"], token)
@@ -34,10 +41,9 @@ class TestCreateAndDeploy(QuestionAnsweringTestCase):
                 "description": "biography of Sir Issac Newton",
                 "language": "en",
                 "multilingualResource": True,
-                "settings": {
-                    "defaultAnswer": "no answer"
-                }
-            })
+                "settings": {"defaultAnswer": "no answer"},
+            },
+        )
 
         # list projects
         qna_projects = client.list_projects()
@@ -58,10 +64,9 @@ class TestCreateAndDeploy(QuestionAnsweringTestCase):
                 "description": "biography of Sir Issac Newton",
                 "language": "en",
                 "multilingualResource": True,
-                "settings": {
-                    "defaultAnswer": "no answer"
-                }
-            })
+                "settings": {"defaultAnswer": "no answer"},
+            },
+        )
 
         # list projects
         qna_projects = client.list_projects()
@@ -86,18 +91,14 @@ class TestCreateAndDeploy(QuestionAnsweringTestCase):
         # test deploy
         deployment_name = "production"
         deployment_poller = client.begin_deploy_project(
-            project_name=project_name,
-            deployment_name=deployment_name,
-            **self.kwargs_for_polling
+            project_name=project_name, deployment_name=deployment_name, **self.kwargs_for_polling
         )
         project = deployment_poller.result()
         assert project["lastDeployedDateTime"]
         assert project["deploymentName"] == "production"
 
         # assert
-        deployments = client.list_deployments(
-            project_name=project_name
-        )
+        deployments = client.list_deployments(project_name=project_name)
         deployment_found = False
         for d in deployments:
             if ("deploymentName" in d) and d["deploymentName"] == deployment_name: